backend/requirements.txt
django==2.1.3
djangorestframework==3.9.0
django-rest-auth==0.9.3
django-allauth==0.38.0
libvirt-python==4.6.0
passlib==1.7.1
paramiko==2.4.1
mysqlclient==1.3.13